oii cheto dolgo delal...

+++ src/base_debayer.h
@@ -6,7 +6,7 @@
 class base_debayer {
 public:
   virtual ~base_debayer() = default;
-  virtual std::vector<uint16_t> debay(const std::vector<uint16_t> &in,
-                                      int width, int height,
-                                      const int debayerOffset[2]) const = 0;
+  virtual std::vector<float> debay(const std::vector<float> &in, int width,
+                                   int height,
+                                   const int debayerOffset[2]) const = 0;
 };
