+++ src/dng_raw.h
@@ -23,8 +23,8 @@ public:
   size_t get_width() override;
   size_t get_height() override;
   int get_white_level() override;
-  std::vector<uint16_t> develop(const base_debayer &debayer,
-                                const class settings &settings) override;
+  std::vector<float> develop(const base_debayer &debayer,
+                             const class settings &settings) override;
 };
 
 #endif // DNG_RAW_H
